Summary: Removing unecessary redundant dependency
Here is an email I sent out describing the issue:


=====================================================



I noticed that you made the following checkin:



revision 1.6
date: 2009/04/09 16:36:48;  author: hkthorn;  state: Exp;  lines: +1 -1
Adding AztecOO and Galeri as optional test libraries.

For some reason building the examples is dependent upon AztecOO and Galeri,
but not the tests even though many of the tests also use Galeri and AztecOO.
When I tried to build Trilinos with Ifpack enabled, but not AztecOO
or Galeri the build failed for the tests.  So, I have added the cmake logic
into the tests that require AztecOO and/or Galeri so these build failures don't
occur.



with the following diff:



Index: Dependencies.cmake
===================================================================
RCS file: /space/CVS/Trilinos/packages/ifpack/cmake/Dependencies.cmake,v
retrieving revision 1.5
retrieving revision 1.6
diff -u -r1.5 -r1.6
--- Dependencies.cmake  27 Nov 2008 23:38:25 -0000      1.5
+++ Dependencies.cmake  9 Apr 2009 16:36:48 -0000       1.6
@@ -1,7 +1,7 @@
 SET(LIB_REQUIRED_DEP_PACKAGES Teuchos Epetra)
 SET(LIB_OPTIONAL_DEP_PACKAGES Amesos AztecOO)
 SET(TEST_REQUIRED_DEP_PACKAGES)
-SET(TEST_OPTIONAL_DEP_PACKAGES Galeri)
+SET(TEST_OPTIONAL_DEP_PACKAGES Galeri AztecOO)
 SET(LIB_REQUIRED_DEP_TPLS)
 SET(LIB_OPTIONAL_DEP_TPLS)
 SET(TEST_REQUIRED_DEP_TPLS)



You added AztecOO as an optional test dependency but it is already an optional library dependency.  You should not have to do this.  In fact, I should have implmented a check to make sure that the same dependency is not repeated in a package like this.

Was there a reason that you changed the file ifpack/cmake/Dependencies.cmake to add this dependency?  Did you just not notice the optional library dependency?


=====================================================


I tested this by bulding Ifpack both with and without AztecOO enabled in
serial debug mode and it worked both ways.