SUMMARY:   Lots of changes to support the StepperValidator initial condition through getPoints check.
     BUGS:      4377

     CHECKLIST:

     1)   Did you add any source files?  YES
     1.a) Did you do a 'cvs add'?  YES
     1.b) Did you add them to a Makefile.am file in some way?  YES
     1.c) Did you do a new bootstrap?  YES

     2)   Did you update or change the build system?  NO

     3) Did you rerun the test suite on at least the packages that you
        changed?  YES

     COMMENTS:

I added an initial condition check to the StepperValidator to make sure you can
get the initial condition out of a stepper after you set the initial condition,
and after you take the first step.  Much of the following changes were required
to accomplish this.

I changed the Backward Euler Stepper, Theta Stepper, and Implicit BDF Stepper
to require an initial condition be set, rather than attempting to get it from
the nominal values on the model.  This required changing the convergence tests
and some of the examples.

I implemented a defaultGetPoints function in StepperHelpers to make it much
easier to handle getPoints from the point of view of a Stepper which doesn't do
its own interpolation, i.e. all the steppers except for Implicit BDF.  This
defaultGetPoints will return initial conditions correctly, and it will use an
interpolator if defined to interpolate between the points.  The following
steppers are converted over to use defaultGetPoints: Explicit RK, Backward
Euler, Forward Euler, Implicit RK, and Theta.  Implicit BDF stepper also uses
it in the post initial condition, pre initialization phase.

I also changed several of the steppers over to a better late initialization
system with the requirement to set an initial condition.

I changed the base interpolator pre-conditions to allow both x=null and
xdot==null, and I added a check to the Hermite interpolator to make sure x !=
null and xdot != null.  And I changed the Linear interpolator to handle
x==null.

I changed the interpolation buffer pre-condidtions to allow xdot==null.

I changed the assumptions on the InterpolationBufferAppenderBase class by
changing the pre-condition check.  Previously, the append range had to sit
directly next to the sink range, which had the consequence that you could not
add [2,3] to [0,1].  The new assumptions are that the append range has to be
entirely on one side or the other, but it doesn't have to match up exactly with
the boundaries of the sink buffer's range.

I changed the order of arguments in selectPointsInTimeRange to put the output
last, and I changed the argument to a Ptr instead of a raw pointer.

I fixed the issue with the PointwiseInterpolationBufferAppender where it would
throw if the append range and the sink range met exactly and shared a point (a
requirement of the previous pre-conditions actually).  Now, it will exclude the
point in common from the append.  I did this by adding a new set of TimeRange
classes which use a different isInRange function observing sets with open and
closed ends.  I'm not entirely happy with the result, but it does work nicely
and it fixes the problem.

I added many tests to the PointwiseInterpolationBufferAppender to explore the
many ways it can operate.

I noticed we didn't have any tests for compareTimeValues, so I added a few, and
I realized that the tolerance for two numbers to be equal is entirely based on
the first argument.  In many of the calls to compareTimeValues in the TimeRange
functions, the same argument is being used with two different calls, and it was
in the second argument.  I swaped the arguments and the sign of the result so
that when two compareTimeValues are being used together, they're using the same
tolerance for the arguments to be equal.

While doing all of this I kept running into a segfault which ended up as bug
#4388, and is being resolved by Ross in Thyra.  There are now several tests
which explore this issue and are #ifdef'd out.

I also added a stub for a global error estimation test, which is currently
turned off due to bug #4388.


(checklist version 2007-06-20)