Cleanning out Intrepid config header files
There was a whole bunch of junk in the Intrepid_config.h.in and
Intrepid_ConfigDefs.hpp helper files that was copied from NewPackage a long
time ago.  The change was to remove all of the optional header check macros.
We decided a *long* time ago (see the trilinos-framwork mail list) that we
were not going to check for the presents of standard C++ headers anymore.

I tested this with GCC 4.2.0 for a serial build and Intrepid build and run
just fine (all 4 tests).