On branch isorropia-triangular-solve: - added CMake support to build new LevelSolver test. - Removed old-timey stand-alone Makefile for this test. - Added optional dependence for Isorropia on ThreadPool, which the test will soon utilize.
The dependence on ThreadPool required that ThreadPool be listed before Isorropia in global TrilinosPackages.cmake.
Implemented this change in the main branch, as the isorropia-triangular-solve should cover only the isoroppia/ directory,
and because it's the right thing to do.

Configured, built and tested Cmake, both with branch and for a fresh checkout of trilinos, for the following packages:
-- Processing enabled package: Teuchos
-- Processing enabled package: ThreadPool
-- Processing enabled package: Epetra
-- Processing enabled package: Zoltan
-- Processing enabled package: Triutils
-- Processing enabled package: EpetraExt
-- Processing enabled package: Isorropia

For fresh:
100% tests passed, 0 tests failed out of 128

For branch (with new test):
100% tests passed, 0 tests failed out of 129