In the LSC strategy and SIMPLE factory some mild improvements and development of "new ideas" (read graph laplacian stuff) was added.
Also in the inverse factory, I added some more refined error handling.
It is somewhat minimal for now, but as we move foreward and see more
errors it will be important for us to figure out what those are. From
that perspective we will see what is important for the error handling
mechanisms in terms of output.

The rest was some housekeeping. Essentially, I removed much of the
early Epetra specific LSC stuff and the corresponding testing which
is now redundant. Also, I removed unneeded functions especially where
the functionality exits elsewhere.  I moved the contents of "PB_Helpers"
into "PB_Utilities" and similarly for "PB_EpetraLSCHelpers". I also
tried to improve the doxygen documentation where I noticed it was lacking.
Though much needs to be done in this regard.