SUMMARY: Implementation of lightweight wrapper class for Hypre parallel IJ matrices.
BUGS: n/a

CHECKLIST:

1) Did you add any source files? yes
1.a) Did you do a 'cvs add'? yes
1.b) Did you add them to the Makefile in some way? yes
1.c) Did you do a new bootstrap? yes

2) Did you update or change the build system? yes
2.a) Did you remember to do a bootstrap and do a 'cvs add' on the Makefiles? yes
2.b) Did you blow away the libraries and executables then rebuild? yes

3.) Did you rerun the test suite on at least the packages that you changed? yes

COMMENTS:

This is the first commit including the EpetraExt_HypreIJMatrix class, a lightweight wrapper for Hypre matrices. It compiles and I haven't found any errors in the implementation. This commit also includes the unit test for the class.

The EpetraExt_HypreIJMatrix class uses Epetra_BasicRowMatrix as its base class, so it is fully functional as an Epetra matrix. I implemented the LeftScale() and RightScale() instead of ExtractMyEntryView() because Hypre doesn't easily offer a view of the matrix.

The makefiles will by default not build Hypre. It is required to have Hypre built on a computer that uses EpetraExt_HypreIJMatrix class.