SUMMARY:   Bugfix for IBDF to support Charon.  Minor changes to the testing for the forward sensitivties.
     CHECKLIST:

     1)   Did you add any source files?  NO

     2)   Did you update or change the build system?  NO

     3) Did you rerun the test suite on at least the packages that you
        changed?  YES

     COMMENTS:

I implemented a defaultGetPoints in the StepperHelpers file to support steppers
that just need to return interpolated values between time-steps.  This function
also supports returning the IC when in a partially initialized state.  I added
this capability to the IBDF stepper incorrectly and my tests didn't catch it,
but Charon's tests diffed.  This is now fixed and I'm in the process of
adding an appropriate test to Rythmos to verify its fixed.

While I was at it, I added an additional nonmember constructor to the IBDF stepper.
I did add a unit test to the IBDF stepper in the hope that I could reproduce
the bug on Charon's side, but my example was too simple and it didn't work.  I
am now going to try a nonlinear example.

The ForwardSensitivity changes are to the unit tests to convert from
TEST_EQUALITY to TEST_FLOATING_EQUALITY.

(checklist version 2007-06-20)