SUMMARY:   Zoltan_LB_Eval now works on a copy of the Zoltan struct,                 rather than on the original, to prevent unintended side                 effects from propagating to the partitioner.  See below                 for more details.
     BUGS:      na

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  no

     COMMENTS:

Zoltan_LB_Eval can call Zoltan_LB_Eval_HG, which uses the same code
to build a hypergraph for Zoltan_LB_Eval_HG as for Zoltan_PHG.  This code
makes some changes to the Zoltan struct; for example, if
LB.Approach=repartition, the hypergraph build code sets LB.Remap=0.  But
if the partitioner is RCB or ParMETIS, we may still want the default value
of LB.Remap=1.  As a result of this side effect, the partitioner gives
different results depending on whether or not Zoltan_LB_Eval is called before
the partitioner.  By having Zoltan_LB_Eval work on a copy of the Zoltan struct,
we avoid any unintended side effects.



(checklist version 2007-06-20)