SUMMARY: A change to how number of neighboring partitions is computed, and a change to how cutn/cutl are computed, in Zoltan_LB_Eval_Graph
     BUGS:   4452

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system? no

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes

     COMMENTS:

In Zoltan_LB_Eval_Graph:

Number of Neighbor Partitions:
  This was computed using a numParts*numParts array before.  Now we use arrays
that are at most numEdges in length, usually much smaller.

CUTN:
  I found an error in the calculation of this for Eval_Graph.  The "hyperedge"
is considered to be the vertex and its neighbors for this calculation.  Then
cutn was calculated using the number of cut edges, instead of the number of
partitions containing the neighbors.

NOTE:
  valgrind reports that 12 bytes are not freed when running zdrive, but it
can't find them.  I am using the libc hsearch() function to keep track of
the number of neighboring partitions.  Maybe it is allocating 12 bytes that
it is not freeing?