Redstorm cross compile files for cmake. This builds the libraries needed for Alegra. I was unable to test as I do not have an account on Redstorm anymore, but will be getting one of the Alegra folks to test it to make sure the build works.
These files were made using the cross compile environment on the scico lan and
are not designed for reddish specifically. however the XC environment on the
scico lan was designed to be very similar to the reddish one so it should work
with very little modification. The changes should really only need to be the
paths in the toolchain file to point to the correct place on reddish.