Made a small change to the Redstorm toolchain file to allow the use of paths not specified in the toolchain file to be used to find libraries and headers. This change allows the cross compile build for Redstorm to use both the host environment as well as the target environment. This was necessary to allow people to use existing tpls that did not follow the directory structure that cmake insists on when you set up the root path for cross compiling.
This does increase the risk of potentially picking up the wrong
libraries/headers, but is the only clean way to allow tpls during cross compiles
right now. The risk should be minor for the Redstorm toolchain though since it
was initially done with disallowing the use of the host environment.