SUMMARY:  Changing the way a "hypergraph" edge weight is assigned to hyperedges created out of graph with no edge weights
     BUGS:  4490

     1)   Did you add any source files?  no
     2)   Did you update or change the build system?  no
     3) Did you rerun the test suite on at least the packages that you
        changed?  some of it

I looked carefully at ch_simple/zdrive.inp.phg on 4 processes.  The calculations
that are being done for cutn and cutl are "correct", but they look odd because of
the choice I had made for hyperedges/hyperedge weights in the ch_simple graph.

Zoltan_LB_Eval_Graph runs BEFORE and AFTER load balancing.

BEFORE, there are no edge weights.  For each of the 25 vertices in the graph,
we take the vertex and it's neighbors to be a "hyperedge", with weight 1.
We do the CUTN and CUTL calculations and got the reasonable looking numbers
22 and 30 for these.

AFTER PHG has run, LB_Eval is called again and now there are edge weights in
the graph - 1.0 per edge.  So now when we do the CUTN and CUTL calculations,
the hyperedge weights are 2, 3 or 4 instead of 1.  And the CUTN,CUTL numbers
are much larger.

I changed Zoltan_LB_Eval_Graph to assign hyperedge weights equal to 1.0 for
each original graph edge weight.  Since this is the default used in PHG, our
BEFORE and AFTER metrics will be comparable