Summary: Fix default generator for non-Experimental
(*) cmake/ctest/TrilinosCTestDriverCore.cmake: Do not try to load
CMakeCache.txt if the build tree is not known.  When the build tree is
already known (in the case of "make Experimental") we can load its cache
and re-use its generator.  Otherwise we fall back to the standard "Unix
Makefiles" default.