Summary: Speedup PACKAGE_COPY_FILES_TO_BINARY_DIR
About 30% of the cmake configuration step when tests are enabled was
being spent repeatedly parsing lists in the list(GET) commands within
the PACKAGE_COPY_FILES_TO_BINARY_DIR function.  This commit improves
configuration time significantly.

(*) cmake/utils/PackageCopyFilesToBinaryDir.cmake: We optimize the case
when the source and destination file list is the same and use foreach()
to iterate through the list once.  This is much faster than using the
list(GET) command which re-parses the list every time (it is meant only
for convenience and small lists).