Getting Rythmos to build without NOX
The recent NOX build failure checked in caused several Rythmos tests to not
build even though NOX is supposed to only be an optional dependency.

This checkin get Rythmos to at least build when NOX is disabled.

I fixed a few of the tests to also work but I don't know if all of the tests
will run.