Provide TEST_CONFIG with PACKAGE_ADD_ADVANCED_TEST
The PACKAGE_ADD_ADVANCED_TEST macro drives its tests through a helper
script DriveAdvancedTest.  In multi-configuration build trees, like
those for Visual Studio on Windows, tests may need to know the name of
the configuration being tested.  We teach PACKAGE_ADD_ADVANCED_TEST to
pass -DTEST_CONFIG=<config> to the test script it generates with the
run-time configuration to be tested.  Then DriveAdvancedTest passes the
configuration on to the test as an environment variable.